Reorganize everything into a namespace
