+++ Source/Actors/Player.cs
@@ -1,5 +1,5 @@
-
 using System.Collections.ObjectModel;
+using Celeste64.Mod;
 
 namespace Celeste64;
