+++ Source/Game.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics;
 using System.Text.Json;
+using Celeste64.Mod;
 
 namespace Celeste64;
