+++ Source/Mod/Helpers/CustomPlayerStateRegistry.cs
@@ -1,4 +1,4 @@
-﻿namespace Celeste64;
+﻿namespace Celeste64.Mod;
 
 public static class CustomPlayerStateRegistry
 {
