+++ Source/Mod/Helpers/SkinInfo.cs
